requirements.txt
fastapi 
uvicorn 
langgraph 
python-dotenv 
langchain-openai